% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinations.R
\docType{data}
\name{combination}
\alias{combination}
\title{Add a combination}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
combination
}
\arguments{
\item{input}{A named list of the form: list(strat1="<some name>")}

\item{output}{A named list of the form list(strat1=c("strategy1", "strategy2", "..."))}
}
\value{
void
}
\description{
Add a combination
}
\keyword{datasets}
