% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{make.budget}
\alias{make.budget}
\title{Generate a list of budgets that adequately tests out different combinations of strategies
Currently computes the prefix sum of the strategy cost vector and mixes it with the strategy costs}
\usage{
make.budget(cost.vector)
}
\arguments{
\item{cost.vector}{A list of numbers}
}
\value{
A sorted list of new budget levels
}
\description{
Generate a list of budgets that adequately tests out different combinations of strategies
Currently computes the prefix sum of the strategy cost vector and mixes it with the strategy costs
}
