% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{opt.result.to.df}
\alias{opt.result.to.df}
\title{TODO: Documentation}
\usage{
opt.result.to.df(opt.result)
}
\arguments{
\item{opt.result}{A list}
}
\value{
slkdfl
}
\description{
TODO: Documentation
}
