% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{Optimize}
\alias{Optimize}
\title{Optimize over a range of budgets and survival thresholds}
\usage{
Optimize(benefits.matrix, cost.vector, all.index = NULL,
  budgets = NULL, thresholds = c(50.01, 60.01, 70.01),
  combo.strategies = NULL, weights = NULL)
}
\arguments{
\item{benefits.matrix}{A [strategies]x[species] dataframe with named rows and columns}

\item{cost.vector}{A list of strategy costs}

\item{all.index}{An integer signifying the index of the strategy that combines all strategies}

\item{budgets}{A list of budgets over which to optimize. If NULL, a sensible range of budgets will be automatically generated}

\item{thresholds}{A list of survival thresholds over which to optimize, default = (50, 60, 70)}

\item{combo.strategies}{A combination object specifying which strategies are combinations of which other strategies}

\item{weights}{A named list of species weights}
}
\value{
A dataframe of optimization results
}
\description{
Optimize over a range of budgets and survival thresholds
}
\examples{
COMBO_info <- get(data("COMBO_info"))
benefits.matrix <- get(data("Bij_fre_01"))
cost.vector <- get(data("cost_fre_01"))$Cost
opt.results <- Optimize(benefits.matrix, cost.vector, combo.strategies=COMBO_info)
PlotResults(opt.results)
}
