% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot.results}
\alias{plot.results}
\title{Plot results of running optimize.range}
\usage{
\method{plot}{results}(summary.results, draw.labels = TRUE)
}
\arguments{
\item{summary.results}{A dataframe containing the results of the complementarity optimization}

\item{draw.labels}{Boolean, indicating whether to draw strategy names on the plot}
}
\value{
A ggplot object
}
\description{
asdasd
}
