% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{PlotResults}
\alias{PlotResults}
\title{Plot results of running optimize.range}
\usage{
PlotResults(summary.results, draw.labels = TRUE)
}
\arguments{
\item{summary.results}{A dataframe containing the results of the complementarity optimization}

\item{draw.labels}{Boolean, indicating whether to draw strategy names on the plot}
}
\value{
this.plot A ggplot object
}
\description{
Provides a convenient way to visualize optimization results over several budget increments and
persistence thresholds.
}
\examples{
res <- Optimize(benefits.matrix, cost.vector, all.index=nrow(benefits.matrix))
PlotResults(res)
}
